# rewrite radix tree