--- a/breaker.go
+++ b/breaker.go
@@ -9,23 +9,16 @@ import (
 	"net/http"
 )
 
-// Breaker is circuit breaker, it's a collection of:
-// - balancer, who response for choose which backend should we proxy, `Host` -> `Balancer` mapper
+// Breaker is circuit breaker, it's a collection of Application
 type Breaker struct {
-	balancers map[string]Balancer
+	balancers map[string]*Application
 }
 
 // NewBreaker return a brand new circuit breaker, with nothing in mapper
 func NewBreaker() *Breaker {
 	return &Breaker{
-		make(map[string]Balancer),
+		make(map[string]*Application),
 	}
-}
-
-// UpdateAPP insert or overwrite a existing app in balancer
-// NOTE: Not concurrency safe!
-func (b *Breaker) UpdateAPP(app string) {
-	b.balancers[app] = NewWRR()
 }
 
 func (b *Breaker) ServeHTTP(w http.ResponseWriter, r *http.Request) {