start to working on it. related: https://github.com/jiajunhuang/guard/issues/6 & https://github.com/jiajunhuang/guard/issues/5