@jiajunhuang 2018-01-23
https://github.com/jiajunhuang/guard/blob/6379f5b774867f3add2c527af6e9119a841853f8/radix_tree.go#L149 need add setMethod