@jiajunhuang 2018-01-24
now, the benchmark is:

```
BenchmarkRdmSelect-4   	50000000	        30.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkRRSelect-4    	100000000	        14.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkWRRSelect-4   	100000000	        20.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkProxy-4       	  200000	      5420 ns/op	   33664 B/op	      12 allocs/op
BenchmarkIncr-4        	200000000	         7.15 ns/op	       0 B/op	       0 allocs/op
BenchmarkQuery-4       	500000000	         3.83 ns/op	       0 B/op	       0 allocs/op
```

before rewrite, benchmark is:

```
BenchmarkRdmSelect-4          	50000000	        32.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkRRSelect-4           	100000000	        14.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkWRRSelect-4          	100000000	        21.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkServeHTTP-4          	  200000	      6042 ns/op	   33712 B/op	      14 allocs/op
BenchmarkProxy-4              	  200000	      5716 ns/op	   33664 B/op	      12 allocs/op
BenchmarkRouterGenericURL-4   	20000000	        96.3 ns/op	      48 B/op	       2 allocs/op
BenchmarkIncr-4               	20000000	        87.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkQueryStatus-4        	100000000	        22.9 ns/op	       0 B/op	       0 allocs/op
```

see? `incr` is 12x faster than old one, `query` is 5.9x faster than old one.