@jiajunhuang 2018-01-24
`time.Now` is really expensive:

```
BenchmarkRdmSelect-4        	50000000	        32.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkRRSelect-4         	100000000	        15.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkWRRSelect-4        	100000000	        23.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkProxy-4            	200000000	         6.25 ns/op	       0 B/op	       0 allocs/op
BenchmarkByPath-4           	300000000	         5.80 ns/op	       0 B/op	       0 allocs/op
BenchmarkByPathNotFound-4   	300000000	         4.70 ns/op	       0 B/op	       0 allocs/op
BenchmarkByPathParam-4      	100000000	        14.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkByPathCatchall-4   	200000000	         7.25 ns/op	       0 B/op	       0 allocs/op
BenchmarkIncr-4             	30000000	        59.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkQuery-4            	20000000	        59.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkRightNow-4         	30000000	        47.0 ns/op	       0 B/op	       0 allocs/op
```