@jiajunhuang 2018-01-24
note that after this rewrite, `ServeHTTP` is 24x faster.

before:

```
BenchmarkRdmSelect-4          	30000000	        47.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkRRSelect-4           	100000000	        17.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkWRRSelect-4          	50000000	        27.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkServeHTTP-4          	  200000	     10462 ns/op	   33712 B/op	      14 allocs/op
BenchmarkProxy-4              	  200000	      9845 ns/op	   33664 B/op	      12 allocs/op
BenchmarkRouterGenericURL-4   	10000000	       155 ns/op	      48 B/op	       2 allocs/op
BenchmarkIncr-4               	20000000	       106 ns/op	       0 B/op	       0 allocs/op
BenchmarkQueryStatus-4        	50000000	        30.9 ns/op	       0 B/op	       0 allocs/op
```

after:

```
BenchmarkRdmSelect-4        	30000000	        47.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkRRSelect-4         	100000000	        16.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkWRRSelect-4        	50000000	        26.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkServeHTTP-4        	 3000000	       428 ns/op	     106 B/op	       3 allocs/op
BenchmarkProxy-4            	  200000	      8820 ns/op	   33664 B/op	      12 allocs/op
BenchmarkByPath-4           	200000000	         8.79 ns/op	       0 B/op	       0 allocs/op
BenchmarkByPathNotFound-4   	200000000	         9.39 ns/op	       0 B/op	       0 allocs/op
BenchmarkByPathParam-4      	50000000	        26.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkByPathCatchall-4   	100000000	        12.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkIncr-4             	20000000	        69.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkQuery-4            	20000000	        68.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkRightNow-4         	30000000	        54.0 ns/op	       0 B/op	       0 allocs/op
```